requirements.txt
numpy >= 1.21.2
pandas >= 1.3.5
scikit-learn==1.0.2

torch==1.10.0
pytorch-ignite==0.4.2
tensorboard==2.11.2
git+https://github.com/VLL-HD/FrEIA.git